chore: move caspi_Envelope.h to control/ subdir
# Conflicts:
#	Tests/CMakeLists.txt
